requirements.txt
numpy>=1.21.0
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
opencv-python>=4.8.0
scikit-learn>=1.0.0
tensorflow>=2.13.0
spacy>=3.6.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.4.0
flake8>=6.1.0
pre-commit>=3.3.3
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.5
sqlalchemy>=2.0.0
alembic>=1.11.0
redis>=4.6.0
celery>=5.3.0 